+++ Controller/WebPayment/ReturnFromWalletGateway.php
@@ -9,7 +9,7 @@ class ReturnFromWalletGateway extends Action
 {
     public function execute()
     {
-        // TODO This is not supposed to be called because wallet is handled by widget
+        // TODO This is not supposed to be called. Wallet is handle by widget
         $resultRaw = $this->resultFactory->create(ResultFactory::TYPE_RAW);
         $resultRaw->setContents('OK');
         return $resultRaw;