+++ Helper/Data.php
@@ -2,7 +2,6 @@
 
 namespace Monext\Payline\Helper;
 
-use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\App\Helper\AbstractHelper;
 use Magento\Framework\App\Helper\Context;
 use Magento\Framework\Math\Random as MathRandom;
@@ -23,6 +22,7 @@ class Data extends AbstractHelper
      * @var Serialize
      */
     protected $serialize;
+
 
     /**
      * @param Context $context
@@ -180,5 +180,21 @@ class Data extends AbstractHelper
         return $amount;
     }
 
+
+    /**
+     * @return string
+     */
+    public function getMerchantName()
+    {
+        $merchantName = $this->scopeConfig->getValue(HelperConstants::CONFIG_PATH_PAYLINE_GENERAL_MERCHANT_NAME,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+
+        if(empty($merchantName)) {
+            $merchantName = $this->scopeConfig->getValue(\Magento\Store\Model\Information::XML_PATH_STORE_INFO_NAME,
+                \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        }
+
+        return  preg_replace('/[^A-Z0-9]/', '', strtoupper($merchantName)) ?? 'UNDEFINEDMERCHANTNAME';
+    }
 
 }